Separate into files and list sessions being managed so that they can be deleted too
